Input logic + Array + Stack
Implemented the logic for taking the input from the user and storing the numbers into an array. Also created a stack and the subroutines to go with it. I wasn't going to make the array or stack yet because I didn't want to get too ahead but it just wasn't going to work if I didn't create and use them now.